Created folder structure and translated files to English
